[M3] refactor: 패키지 경로 정리
참고: https://edu.nextstep.camp/s/DVVUSLFe/ls/LonguG1G
