+++ src/main/java/nextstep/laddergame/enumerate/BridgePositionEnum.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame.domain;
+package nextstep.laddergame.enumerate;
 
 public enum BridgePositionEnum {
     LEFT, RIGHT, NONE
