+++ src/main/java/nextstep/laddergame/enumerate/DirectionEnum.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame.domain;
+package nextstep.laddergame.enumerate;
 
 public enum DirectionEnum {
     RIGHT, LEFT, BOTTOM
