+++ src/main/java/nextstep/laddergame/wrapper/RewardWrapper.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame.domain;
+package nextstep.laddergame.wrapper;
 
 import java.util.List;
