+++ src/test/java/nextstep/laddergame/domain/LadderPieceTest.java
@@ -1,5 +1,6 @@
 package nextstep.laddergame.domain;
 
+import nextstep.laddergame.enumerate.BridgePositionEnum;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
